new AuthLookupMaintenanceJob to check consistency, runs priority 3 on the authlookup queue every 8 hours #1866
