+++ app/jobs/regular_maintenance_job.rb
@@ -20,7 +20,6 @@ class RegularMaintenanceJob < ApplicationJob
     clean_git_repositories
     resend_activation_emails
     remove_unregistered_users
-    check_authlookup_consistency
   end
 
   private
@@ -73,26 +72,4 @@ class RegularMaintenanceJob < ApplicationJob
     end
   end
 
-  # checks lookup_table_consistent? on each type for each user, and if not triggers a job to repopulate for that user
-  # if not already queued
-  def check_authlookup_consistency
-    found_types = [].to_set
-    items_for_queue = [].to_set
-    User.where.not(person_id: nil).to_a.push(nil).each do |user|
-
-      # will only deal with 1 type per user per run
-      found = Seek::Util.authorized_types.find do |type|
-        !type.lookup_table_consistent?(user)
-      end
-
-      next unless found.present?
-
-      found_types << found
-      missing = found.items_missing_from_authlookup(user)
-      items_for_queue.merge(missing)
-    end
-
-    found_types.each(&:remove_invalid_auth_lookup_entries)
-    AuthLookupUpdateQueue.enqueue(items_for_queue.to_a) unless items_for_queue.empty?
-  end
 end
